+++ CHANGELOG.md
@@ -104,6 +104,12 @@
   * Decimals
   * Exceptions
 
+### Changed
+
+- Ensure controls are properly applied and none are missed [#242][242]
+
+[242]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/242
+
 ## [1.22.1][] - 2021-10-04